requirements.txt
pandas
selenium
beautifulsoup4
requests
google-generativeai
openpyxl
lxml
python-dotenv
psutil
py-cpuinfo
undetected-chromedriver
webdriver-manager